requirements.txt
Flask==2.0.2
yowsup==4.0.1
Pillow==8.4.0
requests==2.26.0
pymongo==4.0.1